# web/openapi.yaml
openapi: 3.1.0
info:
  title: Owlette API
  version: 2.3.1
  description: |
    The Owlette API lets you programmatically manage your fleet of Windows machines,
    deploy software, monitor processes, execute tools, and integrate with external systems
    via webhooks.

    ## Authentication

    All API requests require authentication via an **API key**. You can create API keys
    from the Owlette dashboard under **Settings → API Keys**, or via the
    [Create API Key](#tag/API-Keys/operation/createApiKey) endpoint.

    Include your API key in every request using one of these methods:

    - **Header** (recommended): `x-api-key: owk_your_key_here`
    - **Query parameter**: `?api_key=owk_your_key_here`

    API keys are scoped to your user account and inherit your site access permissions.

    ## Rate Limiting

    All endpoints are rate-limited. When you exceed the limit, you'll receive a `429 Too Many Requests`
    response with a `Retry-After` header indicating how long to wait.

    Default limit: **300 requests/hour** per API key.

    Response headers on every request:
    - `X-RateLimit-Limit` — Maximum requests per window
    - `X-RateLimit-Remaining` — Requests remaining in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets

    ## Errors

    All errors return a JSON object with an `error` field:

    ```json
    { "error": "Description of what went wrong" }
    ```

    Common HTTP status codes:
    | Code | Meaning |
...

...

...
  license:
...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /api/roosts/{roostId}/versions/{versionRef}/files:
    x-stub: true
    parameters:
      - $ref: '#/components/parameters/RoostIdPath'
      - $ref: '#/components/parameters/VersionRefPath'
      - $ref: '#/components/parameters/RoostVersionHeader'
    get:
      operationId: listVersionFiles
      tags: [versions]
      summary: paginated flat file list for a version
      description: |
        Flat, cursor-paginated list of files in a version. Avoids
        downloading the full oci layer array for large roosts.
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      parameters:
        - $ref: '#/components/parameters/SiteIdQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageTokenQuery'
        - name: prefix
          in: query
          required: false
          schema:
            type: string
          description: POSIX path prefix filter (e.g. `assets/`).
      responses:
        '200':
          description: Paginated file list.
          content:
            application/json:
              schema:
                type: object
                required: [items, next_page_token]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/VersionFile' }
                  next_page_token:
                    type: string
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }
  /api/roosts/{roostId}/versions/{versionRef}/diff:
    x-stub: true
    parameters:
      - $ref: '#/components/parameters/RoostIdPath'
      - $ref: '#/components/parameters/VersionRefPath'
      - $ref: '#/components/parameters/RoostVersionHeader'
    get:
      operationId: diffVersions
      tags: [versions]
      summary: file-level diff between two versions
      description: |
        Returns the `added`, `removed`, and `modified` file sets
        between the path's version (`to`) and the `against` query
        parameter (`from`).
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      parameters:
        - $ref: '#/components/parameters/SiteIdQuery'
        - name: against
          in: query
          required: true
          schema:
            type: string
            pattern: '^sha256:[0-9a-f]{64}$'
          description: The "from" versionRef to diff against.
      responses:
        '200':
          description: Diff result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiffResult' }
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }

  # ──────────────────────────────────────────────────────────────────────
  # 4. chunks — check/upload-urls/download-urls are live above.
  # mount + referrers are proposed.
  # ──────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: x-api-key
      in: header
      description: |
        Legacy api key header — supported for backwards compatibility.
        New integrations should use `bearerApiKey` instead
        (`Authorization: Bearer owk_live_...`).
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: owk_live_* / owk_test_*
      description: |
        Preferred auth mechanism for the public roost api. Send the
        scoped api key as `Authorization: Bearer owk_live_<secret>`
        (production) or `Authorization: Bearer owk_test_<secret>`
        (sandbox). See [authentication.md](../docs/api/authentication.md).
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: |
        Firebase ID token issued to the authenticated user or agent device.
        Used by dashboard + agent requests to roost endpoints. First-party
        clients only — external integrations use `apiKey` /
        `bearerApiKey` instead.
  parameters:
    SiteIdQuery:
      name: siteId
      in: query
      required: true
      schema:
        type: string
      description: The site ID
...
    RoostIdPath:
      name: roostId
      in: path
      required: true
      schema:
        type: string
      description: The roost ID (deploy unit).
...
    RoostVersionHeader:
      name: Roost-Version
      in: header
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: '2026-04-22'
      description: |
        Selects a dated api version. Optional in wave 1; required once
        a second version ships. See [`GET /api/version`](#tag/utility/operation/getApiVersion).
...
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: Maximum items per page.
    PageTokenQuery:
      name: page_token
      in: query
      required: false
      schema:
        type: string
      description: |
        Opaque cursor from the prior response's `next_page_token`. An
        empty string means no more pages.
...
    VersionRefPath:
      name: versionRef
      in: path
      required: true
      schema:
        type: string
      description: |
        Addresses a version on the roost. Accepts any of:
          - a stable content-addressed id: `vrs_<hex>`
          - a version number: `3`, `#3`, `v3`, or `V3`
          - a roost-relative alias: `current`, `previous`, `first`
        Malformed refs return 400 `version_ref_malformed`;
        unmatched refs return 404 `version_not_found`.
...
  headers:
    RateLimitLimit:
      description: Window limit for the rate-limit bucket that fired for this request.
      schema:
        type: integer
    RateLimitRemaining:
      description: Calls remaining in the current window.
      schema:
        type: integer
    RateLimitReset:
      description: Seconds until the current window fully resets.
      schema:
        type: integer
    RetryAfter:
      description: Seconds the client should wait before retrying (429 only).
      schema:
        type: integer
    RoostRateLimitedReason:
      description: |
        Which rate-limit bucket tripped (`global-rate`, `endpoint-rate`,
        `key-rate`, `site-concurrency`). Present only on 429.
      schema:
        type: string
        enum: [global-rate, endpoint-rate, key-rate, site-concurrency]
  responses:
    Problem400:
      description: Validation failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem401:
      description: Authentication required.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    Problem403:
      description: Forbidden (e.g. cross-site access).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem404:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    Problem429:
      description: Rate limited.
      headers:
        RateLimit-Limit:           { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining:       { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset:           { $ref: '#/components/headers/RateLimitReset' }
        Retry-After:               { $ref: '#/components/headers/RetryAfter' }
        Roost-Rate-Limited-Reason: { $ref: '#/components/headers/RoostRateLimitedReason' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem500:
      description: Internal server error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
...
    ProblemDetails:
      type: object
      description: |
        RFC 7807 problem+json envelope returned by all roost endpoints on
        error. Unlike the legacy `ErrorResponse`, this carries a stable
        machine-readable `type` URI, a correlation `requestId`, and
        optional field-level `errors`.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: Stable URI identifying the problem category.
          example: https://owlette.app/problems/validation-failed
        title:
          type: string
          description: Short, stable human-readable summary.
          example: validation failed
        status:
          type: integer
          description: HTTP status code, mirrors the response status.
          example: 400
        detail:
          type: string
          description: Occurrence-specific explanation; safe to show to users.
        instance:
          type: string
          description: URI reference for this occurrence (often the route path).
        requestId:
          type: string
          description: Correlation ID for trace lookup; quote in support tickets.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: |
            Field-level errors keyed by dotted JSON path (400/422 responses).
            Example: `{ "body.hashes": ["must be a non-empty array"] }`
...
    VersionFile:
      type: object
      required: [path, digest, size]
      properties:
        path:   { type: string }
        digest:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        size:
          type: integer
          format: int64
          minimum: 0
...
    DiffResult:
      type: object
      required: [from, to, added, removed, modified]
      properties:
        from:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        to:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        added:
          type: array
          items: { $ref: '#/components/schemas/VersionFile' }
        removed:
          type: array
          items: { $ref: '#/components/schemas/VersionFile' }
        modified:
          type: array
          items:
            type: object
            required: [path, digestBefore, digestAfter, sizeBefore, sizeAfter]
            properties:
              path: { type: string }
              digestBefore:
                type: string
                pattern: '^sha256:[0-9a-f]{64}$'
              digestAfter:
                type: string
                pattern: '^sha256:[0-9a-f]{64}$'
              sizeBefore:
                type: integer
                format: int64
              sizeAfter:
                type: integer
                format: int64

    # chunks
...